pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "buyeuropean-desktop"
version = "1.0.0"
authors = [
  { name="matu6968", email="wierzejskimateusz8@gmail.com" },
]
description = "Desktop GUI frontend for the BuyEuropean app and it's API using Qt 6 or GTK 4"
license = "MIT"
readme = "README.md"
requires-python = ">=3.9"
keywords = ["BuyEuropean", "GUI", "Qt6", "GTK4"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: End Users/Desktop",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: User Interfaces"
]
dependencies = [
    "requests>=2.28.0",
    "Pillow>=9.0.0"
]

[project.urls]
Homepage = "https://buyeuropean.io"
"Bug Reports" = "https://github.com/matu6968/buyeuropean-desktop/issues"
"Source" = "https://github.com/matu6968/buyeuropean-desktop/"

[project.optional-dependencies]
qt6 = ["PyQt6>=6.0.0"]
gtk4 = ["PyGObject>=3.40.0"]

[project.scripts]
buyeuropean = "buyeuropean.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]
